@@ -23,14 +23,14 @@
   },
   "homepage": "https://github.com/foxdonut/meiosis-examples#readme",
   "dependencies": {
-    "co-body": "^4.0.0",
+    "co-body": "^4.2.0",
     "koa": "^1.2.0",
     "koa-route": "^2.4.2",
     "koa-static": "^2.0.0",
     "serv": "^0.2.4"
   },
   "devDependencies": {
-    "eslint": "^2.8.0",
-    "eslint-plugin-react": "^5.0.1"
+    "eslint": "^2.10.2",
+    "eslint-plugin-react": "^5.1.1"
   }
 }
