@@ -1,5 +1,4 @@
 import h from "snabbdom/h";
-const { div } = require("hyperscript-helpers")(h);
 
 import { Action } from "./actions";
 
@@ -8,24 +7,25 @@ const view = LabeledSlider => (model, actions) => {
   const onRemoveMeasurement = id => actions.sendUpdate(Action.RemoveMeasurement(id));
 
   const renderMeasurement = (measurement, index) =>
-    div({key: measurement.id, style: {border: "1px solid gray"}, id: measurement.id}, [
+    h("div", {key: measurement.id, style: {border: "1px solid gray"}, id: measurement.id}, [
       LabeledSlider({measurement, index}, actions),
-      div([
+      h("div", [
         h("button.btn.btn-danger.btn-sm",
           {on: {click: [onRemoveMeasurement, measurement.id]}}, "Remove Measurement")
       ])
     ]);
 
-  return (
-    div([
-      div(`Measurements: ${model.measurements.map(JSON.stringify)}`),
-      div([
-        h("button.btn.btn-primary.btn-sm",
-          {on: {click: onAddMeasurement}}, "Add Measurement")
-      ]),
-      ...model.measurements.map(renderMeasurement)
+  const measurementNodes = model.measurements.map(renderMeasurement);
+
+  let nodes = [
+    h("div", `Measurements: ${model.measurements.map(JSON.stringify)}`),
+    h("div", [
+      h("button.btn.btn-primary.btn-sm",
+        {on: {click: onAddMeasurement}}, "Add Measurement")
     ])
-  );
+  ].concat(measurementNodes);
+
+  return h("div", nodes);
 };
 
 export default view;
